Sending mail in case of deletion of reserevation

+++ app/Http/Controllers/OptionController.php
@@ -23,6 +23,10 @@ class OptionController extends Controller
 
     public function store(Request $request)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.create')->with('error', ['En tant que fake_admin, vous ne pouvez pas enregistrer de nouvelle option.']);
+        }
+
         $validated = $request->validate([
             'name' => 'required|string',
             'description' => 'nullable|string',
@@ -47,6 +51,10 @@ class OptionController extends Controller
 
     public function update(Request $request, Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.create')->with('error', ['En tans que fake_admin, vous ne pouvez pas modifier d\'option.']);
+        }
+
         $validated = $request->validate([
             'name' => 'required|string',
             'description' => 'required|string',
@@ -64,6 +72,10 @@ class OptionController extends Controller
 
     public function toggleAvailability(Request $request, Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.index')->with('error', ["En tans que fake_admin, vous n'êtes pas autorisé à changer la disponibilité de l'option.<br>Elle correspond à afficher ou non l'option dans la page de réservation."]);
+        }
+
         $option->update([
             'available' => $request->available,
         ]);
@@ -73,6 +85,10 @@ class OptionController extends Controller
 
     public function togglePreselected(Request $request, Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.index')->with('error', ["En tans que fake_admin, vous n'êtes pas autorisé à changer la présélection de l'option.<br>Elle correspond à préselectionner l'option ou non (visuel vert ou orange sur la page de réservation)."]);
+        }
+
         $option->update([
             'preselected' => $request->preselected,
         ]);
@@ -82,6 +98,10 @@ class OptionController extends Controller
 
     public function toggleByDayDisplay(Request $request, Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.index')->with('error', ["En tans que fake_admin... vous n'allez pas toutes les essayer, si ?<br>Affichage de l'input \"par jour ?\" en bas à droite de l'option sur la page de réservation."]);
+        }
+
         $option->update([
             'by_day_display' => $request->by_day_display,
         ]);
@@ -91,6 +111,10 @@ class OptionController extends Controller
 
     public function toggleByDayPreselected(Request $request, Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.index')->with('error', ["En tans que fake_admin, vous n'êtes pas autorisé à changer la présélection de l'option d'option \"Par jour ?\"<br>Cela correspond préselectionner l'option d'option ou non."]);
+        }
+
         $option->update([
             'by_day_preselected' => $request->by_day_preselected,
         ]);
@@ -100,6 +124,10 @@ class OptionController extends Controller
 
     public function destroy(Option $option)
     {
+        if (auth()->user()->role === 'fake_admin') {
+            return redirect()->route('admin.options.index')->with('error', ["Comment osez-vous cliquer sur ce bouton fake_admin ?! ;)"]);
+        }
+
         $option->delete();
         return redirect()->route('admin.options.index')->with('success', ['Option supprimée.']);
     }
